# Get maximum and minimum pebble db version with a single lock call