In all but one case, we need both versions. Getting them together slightly speeds up some operations.

Benchmark results:
```
                                  │     base     │               minmax               │
                                  │    sec/op    │   sec/op     vs base               │
Set/LocalPebble10/AC-24             10.53µ ±  3%   10.55µ ± 2%       ~ (p=0.606 n=11)
Set/LocalPebble10/CAS-24            8.877µ ±  4%   8.697µ ± 2%       ~ (p=0.058 n=11)
Set/LocalPebble100/AC-24            15.43µ ±  4%   15.32µ ± 2%       ~ (p=0.898 n=11)
Set/LocalPebble100/CAS-24           14.09µ ±  6%   13.24µ ± 3%  -6.01% (p=0.004 n=11)
Set/LocalPebble1000/AC-24           25.77µ ± 12%   23.58µ ± 3%       ~ (p=0.056 n=11)
Set/LocalPebble1000/CAS-24          21.77µ ±  5%   21.52µ ± 5%       ~ (p=0.606 n=11)
Set/LocalPebble10000/AC-24          143.3µ ±  5%   139.9µ ± 1%       ~ (p=0.478 n=11)
Set/LocalPebble10000/CAS-24         141.5µ ±  4%   138.8µ ± 2%       ~ (p=0.065 n=11)
Set/LocalPebble1000000/AC-24        3.183m ±  1%   3.195m ± 0%       ~ (p=0.847 n=11)
Set/LocalPebble1000000/CAS-24       3.211m ±  1%   3.177m ± 0%  -1.06% (p=0.001 n=11)
Read/LocalPebble10-24               4.078µ ±  1%   4.060µ ± 1%       ~ (p=0.340 n=11)
Read/LocalPebble100-24              4.636µ ±  3%   4.573µ ± 1%       ~ (p=0.120 n=11)
Read/LocalPebble1000-24             5.824µ ±  2%   5.749µ ± 1%  -1.29% (p=0.020 n=11)
Read/LocalPebble10000-24            23.00µ ±  1%   22.78µ ± 1%       ~ (p=0.065 n=11)
GetSingle/LocalPebble10/AC-24       5.950µ ±  2%   5.885µ ± 2%       ~ (p=0.151 n=11)
GetSingle/LocalPebble10/CAS-24      4.333µ ±  2%   4.304µ ± 1%       ~ (p=0.116 n=11)
GetSingle/LocalPebble100/AC-24      6.285µ ±  3%   6.355µ ± 2%       ~ (p=0.449 n=11)
GetSingle/LocalPebble100/CAS-24     4.726µ ±  2%   4.701µ ± 2%       ~ (p=0.859 n=11)
GetSingle/LocalPebble1000/AC-24     7.210µ ±  4%   7.271µ ± 6%       ~ (p=0.699 n=11)
GetSingle/LocalPebble1000/CAS-24    5.624µ ±  1%   5.667µ ± 1%       ~ (p=0.101 n=11)
GetSingle/LocalPebble10000/AC-24    23.18µ ±  7%   22.94µ ± 6%       ~ (p=0.270 n=11)
GetSingle/LocalPebble10000/CAS-24   21.93µ ±  1%   21.57µ ± 1%  -1.63% (p=0.016 n=11)
GetMulti/LocalPebble10-24           260.1µ ±  2%   259.0µ ± 1%       ~ (p=0.478 n=11)
GetMulti/LocalPebble100-24          289.7µ ±  6%   286.8µ ± 1%       ~ (p=0.171 n=11)
GetMulti/LocalPebble1000-24         392.5µ ±  3%   385.1µ ± 2%       ~ (p=0.133 n=11)
GetMulti/LocalPebble10000-24        1.988m ±  3%   1.910m ± 0%  -3.90% (p=0.000 n=11)
FindMissing/LocalPebble10-24        202.5µ ±  2%   200.4µ ± 2%       ~ (p=0.133 n=11)
FindMissing/LocalPebble100-24       205.8µ ±  1%   203.4µ ± 1%  -1.16% (p=0.023 n=11)
FindMissing/LocalPebble1000-24      205.2µ ±  2%   202.3µ ± 3%       ~ (p=0.300 n=11)
FindMissing/LocalPebble10000-24     202.6µ ±  2%   201.7µ ± 2%       ~ (p=0.270 n=11)
geomean                             42.93µ         42.34µ       -1.38%

                                  │     base      │               minmax                │
                                  │      B/s      │     B/s       vs base               │
Set/LocalPebble10/AC-24             927.7Ki ±  2%   927.7Ki ± 5%       ~ (p=0.734 n=11)
Set/LocalPebble10/CAS-24            1.078Mi ±  4%   1.097Mi ± 2%  +1.77% (p=0.042 n=11)
Set/LocalPebble100/AC-24            6.180Mi ±  3%   6.227Mi ± 2%       ~ (p=0.860 n=11)
Set/LocalPebble100/CAS-24           6.771Mi ± 10%   7.200Mi ± 4%  +6.34% (p=0.004 n=11)
Set/LocalPebble1000/AC-24           37.00Mi ± 10%   40.45Mi ± 2%       ~ (p=0.056 n=11)
Set/LocalPebble1000/CAS-24          43.81Mi ± 12%   44.32Mi ± 5%       ~ (p=0.606 n=11)
Set/LocalPebble10000/AC-24          66.57Mi ±  5%   68.17Mi ± 1%       ~ (p=0.478 n=11)
Set/LocalPebble10000/CAS-24         67.38Mi ±  4%   68.73Mi ± 2%       ~ (p=0.065 n=11)
Set/LocalPebble1000000/AC-24        299.6Mi ±  3%   298.5Mi ± 0%       ~ (p=0.847 n=11)
Set/LocalPebble1000000/CAS-24       297.0Mi ±  2%   300.2Mi ± 0%  +1.07% (p=0.001 n=11)
Read/LocalPebble10-24               2.337Mi ±  2%   2.346Mi ± 1%       ~ (p=0.254 n=11)
Read/LocalPebble100-24              20.57Mi ±  2%   20.86Mi ± 2%       ~ (p=0.120 n=11)
Read/LocalPebble1000-24             163.7Mi ±  2%   165.9Mi ± 1%  +1.30% (p=0.023 n=11)
Read/LocalPebble10000-24            414.6Mi ±  1%   418.7Mi ± 1%       ~ (p=0.065 n=11)
GetSingle/LocalPebble10/AC-24       1.602Mi ±  2%   1.621Mi ± 1%       ~ (p=0.126 n=11)
GetSingle/LocalPebble10/CAS-24      2.203Mi ±  2%   2.213Mi ± 1%       ~ (p=0.117 n=11)
GetSingle/LocalPebble100/AC-24      15.17Mi ±  4%   15.01Mi ± 2%       ~ (p=0.449 n=11)
GetSingle/LocalPebble100/CAS-24     20.18Mi ±  2%   20.28Mi ± 2%       ~ (p=0.859 n=11)
GetSingle/LocalPebble1000/AC-24     132.3Mi ±  4%   131.2Mi ± 6%       ~ (p=0.699 n=11)
GetSingle/LocalPebble1000/CAS-24    169.6Mi ±  1%   168.3Mi ± 1%       ~ (p=0.101 n=11)
GetSingle/LocalPebble10000/AC-24    411.4Mi ±  7%   415.7Mi ± 6%       ~ (p=0.270 n=11)
GetSingle/LocalPebble10000/CAS-24   434.9Mi ±  3%   442.1Mi ± 1%  +1.65% (p=0.016 n=11)
GetMulti/LocalPebble10-24           3.662Mi ±  2%   3.681Mi ± 1%       ~ (p=0.466 n=11)
GetMulti/LocalPebble100-24          32.92Mi ±  4%   33.25Mi ± 1%       ~ (p=0.166 n=11)
GetMulti/LocalPebble1000-24         243.0Mi ±  4%   247.7Mi ± 3%       ~ (p=0.133 n=11)
GetMulti/LocalPebble10000-24        479.8Mi ±  3%   499.3Mi ± 0%  +4.06% (p=0.000 n=11)
geomean                             33.71Mi         34.20Mi       +1.45%

                                  │     base      │                minmax                │
                                  │     B/op      │     B/op       vs base               │
Set/LocalPebble10/AC-24             6.979Ki ±  0%   6.981Ki ±  0%       ~ (p=0.571 n=11)
Set/LocalPebble10/CAS-24            5.104Ki ±  0%   5.103Ki ±  0%       ~ (p=0.476 n=11)
Set/LocalPebble100/AC-24            7.104Ki ±  0%   7.104Ki ±  0%       ~ (p=0.637 n=11)
Set/LocalPebble100/CAS-24           5.229Ki ±  0%   5.229Ki ±  0%       ~ (p=0.571 n=11)
Set/LocalPebble1000/AC-24           8.643Ki ±  0%   8.645Ki ±  0%       ~ (p=0.410 n=11)
Set/LocalPebble1000/CAS-24          6.758Ki ±  0%   6.765Ki ±  0%       ~ (p=0.936 n=11)
Set/LocalPebble10000/AC-24          9.675Ki ±  0%   9.668Ki ±  0%       ~ (p=0.529 n=11)
Set/LocalPebble10000/CAS-24         7.794Ki ±  0%   7.790Ki ±  0%       ~ (p=0.356 n=11)
Set/LocalPebble1000000/AC-24        13.11Ki ± 21%   15.61Ki ± 35%       ~ (p=0.949 n=11)
Set/LocalPebble1000000/CAS-24       11.18Ki ± 24%   11.40Ki ± 26%       ~ (p=0.797 n=11)
Read/LocalPebble10-24               2.966Ki ±  0%   2.966Ki ±  0%       ~ (p=0.495 n=11)
Read/LocalPebble100-24              3.300Ki ±  1%   3.292Ki ±  0%       ~ (p=0.307 n=11)
Read/LocalPebble1000-24             6.719Ki ±  3%   6.593Ki ±  1%       ~ (p=0.068 n=11)
Read/LocalPebble10000-24            32.16Ki ±  3%   32.02Ki ±  2%       ~ (p=0.884 n=11)
GetSingle/LocalPebble10/AC-24       8.058Ki ±  0%   8.058Ki ±  0%       ~ (p=0.820 n=11)
GetSingle/LocalPebble10/CAS-24      2.993Ki ±  0%   2.993Ki ±  0%       ~ (p=1.000 n=11)
GetSingle/LocalPebble100/AC-24      8.100Ki ±  0%   8.100Ki ±  0%       ~ (p=0.665 n=11)
GetSingle/LocalPebble100/CAS-24     3.129Ki ±  0%   3.129Ki ±  0%       ~ (p=0.958 n=11)
GetSingle/LocalPebble1000/AC-24     8.409Ki ±  0%   8.409Ki ±  0%       ~ (p=0.711 n=11)
GetSingle/LocalPebble1000/CAS-24    4.332Ki ±  0%   4.332Ki ±  0%       ~ (p=0.858 n=11)
GetSingle/LocalPebble10000/AC-24    18.39Ki ±  0%   18.39Ki ±  0%       ~ (p=0.180 n=11)
GetSingle/LocalPebble10000/CAS-24   13.31Ki ±  0%   13.31Ki ±  0%       ~ (p=0.545 n=11)
GetMulti/LocalPebble10-24           217.1Ki ±  0%   217.1Ki ±  0%       ~ (p=0.936 n=11)
GetMulti/LocalPebble100-24          230.8Ki ±  0%   230.8Ki ±  0%       ~ (p=0.748 n=11)
GetMulti/LocalPebble1000-24         350.9Ki ±  0%   350.9Ki ±  0%       ~ (p=0.949 n=11)
GetMulti/LocalPebble10000-24        1.219Mi ±  0%   1.219Mi ±  0%       ~ (p=0.748 n=11)
FindMissing/LocalPebble10-24        178.9Ki ±  0%   178.9Ki ±  0%       ~ (p=0.797 n=11)
FindMissing/LocalPebble100-24       183.1Ki ±  0%   183.1Ki ±  0%       ~ (p=0.962 n=11)
FindMissing/LocalPebble1000-24      214.4Ki ±  0%   214.4Ki ±  0%       ~ (p=0.898 n=11)
FindMissing/LocalPebble10000-24     184.4Ki ±  0%   184.4Ki ±  0%       ~ (p=0.468 n=11)
geomean                             19.12Ki         19.23Ki        +0.57%

                                  │    base     │                minmax                │
                                  │  allocs/op  │  allocs/op   vs base                 │
Set/LocalPebble10/AC-24              99.00 ± 0%    99.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble10/CAS-24             72.00 ± 0%    72.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble100/AC-24             97.00 ± 0%    97.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble100/CAS-24            70.00 ± 0%    70.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble1000/AC-24            97.00 ± 0%    97.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble1000/CAS-24           70.00 ± 0%    70.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble10000/AC-24           124.0 ± 0%    124.0 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble10000/CAS-24          97.00 ± 0%    97.00 ± 0%       ~ (p=1.000 n=11) ¹
Set/LocalPebble1000000/AC-24         128.0 ± 2%    129.0 ± 1%       ~ (p=0.742 n=11)
Set/LocalPebble1000000/CAS-24        103.0 ± 2%    103.0 ± 2%       ~ (p=0.975 n=11)
Read/LocalPebble10-24                51.00 ± 0%    51.00 ± 0%       ~ (p=1.000 n=11) ¹
Read/LocalPebble100-24               51.00 ± 0%    51.00 ± 0%       ~ (p=1.000 n=11) ¹
Read/LocalPebble1000-24              51.00 ± 0%    51.00 ± 0%       ~ (p=1.000 n=11) ¹
Read/LocalPebble10000-24             52.00 ± 0%    52.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble10/AC-24        68.00 ± 0%    68.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble10/CAS-24       53.00 ± 0%    53.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble100/AC-24       68.00 ± 0%    68.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble100/CAS-24      53.00 ± 0%    53.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble1000/AC-24      68.00 ± 0%    68.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble1000/CAS-24     53.00 ± 0%    53.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble10000/AC-24     70.00 ± 0%    70.00 ± 0%       ~ (p=1.000 n=11) ¹
GetSingle/LocalPebble10000/CAS-24    54.00 ± 0%    54.00 ± 0%       ~ (p=1.000 n=11) ¹
GetMulti/LocalPebble10-24           3.412k ± 0%   3.412k ± 0%       ~ (p=0.931 n=11)
GetMulti/LocalPebble100-24          3.412k ± 0%   3.413k ± 0%       ~ (p=1.000 n=11)
GetMulti/LocalPebble1000-24         3.413k ± 0%   3.413k ± 0%       ~ (p=1.000 n=11)
GetMulti/LocalPebble10000-24        3.518k ± 0%   3.518k ± 0%       ~ (p=0.237 n=11)
FindMissing/LocalPebble10-24        2.610k ± 0%   2.610k ± 0%       ~ (p=1.000 n=11)
FindMissing/LocalPebble100-24       2.610k ± 0%   2.610k ± 0%       ~ (p=1.000 n=11) ¹
FindMissing/LocalPebble1000-24      2.610k ± 0%   2.610k ± 0%       ~ (p=1.000 n=11)
FindMissing/LocalPebble10000-24     2.610k ± 0%   2.610k ± 0%       ~ (p=1.000 n=11) ¹
geomean                              193.7         193.8       +0.03%
```